fix: NoArgs 제거 & 멤버에 final 추가 & 테스트를 위해 멤버변수 추가
