+++ src/main/java/com/spectrum/service/dto/PostCreateDto.java
@@ -1,16 +1,13 @@
 package com.spectrum.service.dto;
 
 import com.spectrum.domain.post.Post;
-import lombok.AccessLevel;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class PostCreateDto {
-    private String title;
-    private String content;
-    private Long memberId;
+    private final String title;
+    private final String content;
+    private final Long memberId;
 
     public PostCreateDto(String title, String content, Long memberId) {
         this.title = title;
